feat: exit on loop (#101)
## Summary

This PR add a way to detect if we are stuck in a loop. Sometimes node
don't have the rest of the chain so we need to connect to another node.
We added a check to verify if the blocks headers return is either empty
or the same as our current one.

In addition there is a check that compare the current time with the
block to timestamp to determine if the block has been recently created.
In the case the block has been created during the last 15 seconds. We
wait 15 seconds so the next block can be created.

## What changed

* A check to verify we are not stuck in a loop
* Determine if the block has been created in the last 15 seconds,
meaning we have the lastest block and we can wait 15 seconds that's the
next one is created